+++ packages/cli/CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Add new command `codegen:generate` (#1896)
+- Add new command `codegen:generate`, allowing users to generate a scaffold based on selected ABIs (#1896)
+- `subql init` supports scaffold creation via `codegen:generate` (#1896)
 - Support for update availability logs (#1930)
 
 ### Changed